HareRestful add getting yarn running application name in the JobQueueResource class

+++ HareDB_HBaseClient_FacadeAPI/src/main/java/com/haredb/harespark/operator/HareSparkOperator.java
@@ -3,7 +3,7 @@ package com.haredb.harespark.operator;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
-
+import com.haredb.client.facade.bean.MessageInfo;
 import com.haredb.client.facade.until.HareSparkSysConfig;
 import com.haredb.harespark.bean.TableColumn;
 import com.haredb.harespark.bean.TableInfo;
@@ -36,7 +36,7 @@ public class HareSparkOperator {
 	private SysConfig sysConfig;
 	private UserSessionBean userSessionBean;
 	private HareSparkFacade hareSparkFacade;
-	
+	public final static String USERSESSIONNULLMSG = "usersession is null, please connect to usersession restful";
 	public HareSparkOperator() {
 	
 	}
@@ -54,9 +54,9 @@ public class HareSparkOperator {
 		String folder = userSessionBean.getConfigurationFolderPath();
 		File file = new File(folder);
 		if (file.exists()) {
-			bean.setStatus(bean.SUCCESS);
+			bean.setStatus(MessageInfo.SUCCESS);
 		}else {
-			bean.setStatus(bean.ERROR);
+			bean.setStatus(MessageInfo.ERROR);
 			bean.setException("Folder path '" + folder +"' is not exist. " );
 		}
 		stopTime = System.currentTimeMillis();
@@ -70,11 +70,18 @@ public class HareSparkOperator {
 		
 		startTime = System.currentTimeMillis();
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			hareSparkFacade.createTable(createTableBean.getTablename(), createTableBean.getColumnNames(), createTableBean.getDataTypes());
-			bean.setStatus(bean.SUCCESS);
+
+			if(userSessionBean == null){
+				bean.setStatus(MessageInfo.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				hareSparkFacade.createTable(createTableBean.getTablename(), createTableBean.getColumnNames(), createTableBean.getDataTypes());
+				bean.setStatus(MessageInfo.SUCCESS);
+			}
+			
 		}catch(Exception e) {
-			bean.setStatus(bean.ERROR);
+			bean.setStatus(MessageInfo.ERROR);
 			bean.setException("Create table "+ createTableBean.getTablename() + " failed : " + e.getMessage());
 			e.printStackTrace();
 		}		
@@ -87,22 +94,27 @@ public class HareSparkOperator {
 		UploadDataFileResponseBean bean = new UploadDataFileResponseBean();
 		startTime = System.currentTimeMillis();		
 		// Todo : change to thread	
-		try {			
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			boolean skipHeader = false;
-			try {
-				skipHeader = Boolean.valueOf(uploadDataFileBean.getSkipHeader());
-			}catch (Exception e) {
-				throw new Exception("skipHeader value " + uploadDataFileBean.getSkipHeader() + " is error");
-			}			
-			asychronizeRunnable runnable = new asychronizeRunnable(hareSparkFacade, uploadDataFileBean);
-			new Thread(runnable).start();			
-//			hareSparkFacade.uploadDataFile(uploadDataFileBean.getDataFilePath(), uploadDataFileBean.getResultPath(), uploadDataFileBean.getCsvSeparator(), uploadDataFileBean.getTableName(), skipHeader, uploadDataFileBean.getOperator());
-			
-			bean.setUploadJobName(uploadDataFileBean.getResultPath());
-			bean.setStatus(bean.SUCCESS);
+		try {		
+			if(userSessionBean == null){
+				bean.setStatus(MessageInfo.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				boolean skipHeader = false;
+				try {
+					skipHeader = Boolean.valueOf(uploadDataFileBean.getSkipHeader());
+				}catch (Exception e) {
+					throw new Exception("skipHeader value " + uploadDataFileBean.getSkipHeader() + " is error");
+				}			
+				asychronizeRunnable runnable = new asychronizeRunnable(hareSparkFacade, uploadDataFileBean);
+				new Thread(runnable).start();			
+	//			hareSparkFacade.uploadDataFile(uploadDataFileBean.getDataFilePath(), uploadDataFileBean.getResultPath(), uploadDataFileBean.getCsvSeparator(), uploadDataFileBean.getTableName(), skipHeader, uploadDataFileBean.getOperator());
+				
+				bean.setUploadJobName(uploadDataFileBean.getResultPath());
+				bean.setStatus(MessageInfo.SUCCESS);
+			}
 		}catch(Exception e) {
-			bean.setStatus(bean.ERROR);
+			bean.setStatus(MessageInfo.ERROR);
 			bean.setException("Upload DataFile failed : " + e.getMessage());
 			e.printStackTrace();
 		}	
@@ -116,16 +128,21 @@ public class HareSparkOperator {
 		UploadDataFileStatusResponseBean bean = new UploadDataFileStatusResponseBean();
 		startTime = System.currentTimeMillis();		
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			String response = hareSparkFacade.uploadFileStatus(uploadDataFileStatusBean.getUploadJobName());
-			
-			String[] responseinfo = response.split(",");
-			bean.setStartTime(responseinfo[0]);
-			bean.setFinishTime(responseinfo[1]);
-			bean.setDataStatus(responseinfo[2]);
-			bean.setStatus(bean.SUCCESS);
+			if(userSessionBean == null){
+				bean.setStatus(MessageInfo.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				String response = hareSparkFacade.uploadFileStatus(uploadDataFileStatusBean.getUploadJobName());
+				
+				String[] responseinfo = response.split(",");
+				bean.setStartTime(responseinfo[0]);
+				bean.setFinishTime(responseinfo[1]);
+				bean.setDataStatus(responseinfo[2]);
+				bean.setStatus(MessageInfo.SUCCESS);
+			}
 		}catch(Exception e) {
-			bean.setStatus(bean.ERROR);
+			bean.setStatus(MessageInfo.ERROR);
 			bean.setException("Query DataFile Status failed : " + e.getMessage());
 			e.printStackTrace();
 		}	
@@ -140,16 +157,19 @@ public class HareSparkOperator {
 		startTime = System.currentTimeMillis();		
 		try {
 			String jobname = querySubmitBean.getTablename() + "_" + System.currentTimeMillis();
-			
-			
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			asychronizeRunnable runnable = new asychronizeRunnable(hareSparkFacade, querySubmitBean, jobname);
-			new Thread(runnable).start();
-//			hareSparkFacade.queryHDFSTable(querySubmitBean.getResultFileFolder(), querySubmitBean.getSql(), querySubmitBean.getTablename(), jobname);
-			bean.setQueryJobName(jobname);
-			bean.setStatus(bean.SUCCESS);
+			if(userSessionBean == null){
+				bean.setStatus(MessageInfo.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				asychronizeRunnable runnable = new asychronizeRunnable(hareSparkFacade, querySubmitBean, jobname);
+				new Thread(runnable).start();
+	//			hareSparkFacade.queryHDFSTable(querySubmitBean.getResultFileFolder(), querySubmitBean.getSql(), querySubmitBean.getTablename(), jobname);
+				bean.setQueryJobName(jobname);
+				bean.setStatus(MessageInfo.SUCCESS);
+			}
 		}catch(Exception e) {
-			bean.setStatus(bean.ERROR);
+			bean.setStatus(MessageInfo.ERROR);
 			bean.setException("QuerySubmit failed : " + e.getMessage());
 			e.printStackTrace();
 		}	
@@ -164,18 +184,27 @@ public class HareSparkOperator {
 		QueryStatusResponseBean bean = new QueryStatusResponseBean();
 		startTime = System.currentTimeMillis();		
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);			
-			List<String> responselist = hareSparkFacade.queryHDFSTableStatus(queryStatusBean.getQueryJobName());
-			bean.setJobStatus(responselist.get(0));
-			bean.setJobID(responselist.get(1));
-			bean.setJobName(responselist.get(2));
-			bean.setJobFinishTime(responselist.get(3));
-			bean.setJobStartTime(responselist.get(4));
-			bean.setJobProgress(responselist.get(5));
-			
-			bean.setStatus(bean.SUCCESS);
+			if(userSessionBean == null){
+				bean.setStatus(MessageInfo.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);			
+				List<String> responselist = hareSparkFacade.queryHDFSTableStatus(queryStatusBean.getQueryJobName());
+				if(responselist != null){	
+					bean.setJobStatus(responselist.get(0));
+					bean.setJobID(responselist.get(1));
+					bean.setJobName(responselist.get(2));
+					bean.setJobFinishTime(responselist.get(3));
+					bean.setJobStartTime(responselist.get(4));
+					bean.setJobProgress(responselist.get(5));
+					bean.setStatus(MessageInfo.SUCCESS);
+				}else{
+					bean.setStatus(MessageInfo.ERROR);
+					bean.setException("QueryStatus failed : This is JobID not running");
+				}
+			}
 		}catch(Exception e) {
-			bean.setStatus(bean.ERROR);
+			bean.setStatus(MessageInfo.ERROR);
 			bean.setException("QueryStatus failed : " + e.getMessage());
 			e.printStackTrace();
 		}
@@ -190,40 +219,43 @@ public class HareSparkOperator {
 		PreviewResponseBean bean = new PreviewResponseBean();
 		startTime = System.currentTimeMillis();		
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			int pagesize = 0;
-			int limit = 0;
-			
-			try {
-				pagesize = Integer.valueOf(previewBean.getPageSize());
-				limit = Integer.valueOf(previewBean.getLimit());
-			}catch (Exception e) {
-				throw new Exception(e.getMessage());
-			}			
-			List<String> responselist = hareSparkFacade.previewData(previewBean.getTablename(), pagesize, limit);				
-			if (responselist == null || responselist.size() == 0) {
-				throw new Exception("table " + previewBean.getTablename() + " is not created");
-			}		
-			
-			String[] header = responselist.get(0).split(",");
-			
-			String[][] results = new String[responselist.size() - 1][];
-			if (responselist.size() != 1) {
-				for(int i = 1; i < responselist.size(); i++) {
-					String[] temp = responselist.get(i).split(",");
-					results[i - 1] = new String[temp.length];
-					for (int j = 0; j < temp.length; j++) {
-						results[i - 1][j] = temp[j];
-					}					
+			if(userSessionBean == null){
+				bean.setStatus(MessageInfo.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				int pagesize = 0;
+				int limit = 0;
+				
+				try {
+					pagesize = Integer.valueOf(previewBean.getPageSize());
+					limit = Integer.valueOf(previewBean.getLimit());
+				}catch (Exception e) {
+					throw new Exception(e.getMessage());
+				}			
+				List<String> responselist = hareSparkFacade.previewData(previewBean.getTablename(), pagesize, limit);				
+				if (responselist == null || responselist.size() == 0) {
+					throw new Exception("table " + previewBean.getTablename() + " is not created");
+				}		
+				
+				String[] header = responselist.get(0).split(",");
+				
+				String[][] results = new String[responselist.size() - 1][];
+				if (responselist.size() != 1) {
+					for(int i = 1; i < responselist.size(); i++) {
+						String[] temp = responselist.get(i).split(",");
+						results[i - 1] = new String[temp.length];
+						for (int j = 0; j < temp.length; j++) {
+							results[i - 1][j] = temp[j];
+						}					
+					}
 				}
+				bean.setHeads(header);
+				bean.setResults(results);
+				bean.setStatus(MessageInfo.SUCCESS);
 			}
-			
-			
-			bean.setHeads(header);
-			bean.setResults(results);
-			bean.setStatus(bean.SUCCESS);
 		}catch(Exception e) {
-			bean.setStatus(bean.ERROR);
+			bean.setStatus(MessageInfo.ERROR);
 			bean.setException("QuerySubmit failed : " + e.getMessage());
 			e.printStackTrace();
 		}
@@ -238,11 +270,17 @@ public class HareSparkOperator {
 		startTime = System.currentTimeMillis();		
 
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			boolean result = hareSparkFacade.deleteDataFile(deleteDataFileBean.getTablename(), deleteDataFileBean.getDeleteDataFileName());			
-			bean.setStatus(bean.SUCCESS);			
+			if(userSessionBean == null){
+				bean.setStatus(MessageInfo.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				boolean result = hareSparkFacade.deleteDataFile(deleteDataFileBean.getTablename(), deleteDataFileBean.getDeleteDataFileName());			
+				bean.setStatus(MessageInfo.SUCCESS);
+			}
+			
 		}catch(Exception e) {
-			bean.setStatus(bean.ERROR);
+			bean.setStatus(MessageInfo.ERROR);
 			bean.setException("Delete DataFile failed : " + e.getMessage());
 			e.printStackTrace();
 		}	
@@ -257,11 +295,16 @@ public class HareSparkOperator {
 		startTime = System.currentTimeMillis();		
 
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			hareSparkFacade.dropTable(dropTableBean.getTablename());
-			bean.setStatus(bean.SUCCESS);
+			if(userSessionBean == null){
+				bean.setStatus(MessageInfo.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				hareSparkFacade.dropTable(dropTableBean.getTablename());
+				bean.setStatus(MessageInfo.SUCCESS);
+			}
 		}catch(Exception e) {
-			bean.setStatus(bean.ERROR);
+			bean.setStatus(MessageInfo.ERROR);
 			bean.setException("Drop table " + dropTableBean.getTablename() + " failed : " + e.getMessage());
 			e.printStackTrace();
 		}	
@@ -275,11 +318,16 @@ public class HareSparkOperator {
 		ResponseInfoBean bean = new ResponseInfoBean();
 		startTime = System.currentTimeMillis();		
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			hareSparkFacade.alterTable(alterTableBean.getTablename(), alterTableBean.getColumnNames(), alterTableBean.getDataTypes());
-			bean.setStatus(bean.SUCCESS);
+			if(userSessionBean == null){
+				bean.setStatus(MessageInfo.ERROR);
+				bean.setException(USERSESSIONNULLMSG);	
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				hareSparkFacade.alterTable(alterTableBean.getTablename(), alterTableBean.getColumnNames(), alterTableBean.getDataTypes());
+				bean.setStatus(MessageInfo.SUCCESS);
+			}
 		}catch(Exception e) {
-			bean.setStatus(bean.ERROR);
+			bean.setStatus(MessageInfo.ERROR);
 			bean.setException("Alter table " + alterTableBean.getTablename() +" failed : " + e.getMessage());
 			e.printStackTrace();
 		}	
@@ -293,25 +341,30 @@ public class HareSparkOperator {
 		DescribeTableResponseBean bean = new DescribeTableResponseBean();
 		startTime = System.currentTimeMillis();		
 		try {
-			hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
-			TableInfo tableinfo = hareSparkFacade.describeTable(describeTableBean.getTablename());			
-			
-			if (tableinfo == null) {
-				throw new Exception("Table " + describeTableBean.getTablename() + " is not existed");
+			if(userSessionBean == null){
+				bean.setStatus(MessageInfo.ERROR);
+				bean.setException(USERSESSIONNULLMSG);
+			}else{
+				hareSparkFacade = new HareSparkFacade(userSessionBean.getConfigurationFolderPath(), sysConfig);
+				TableInfo tableinfo = hareSparkFacade.describeTable(describeTableBean.getTablename());			
+				
+				if (tableinfo == null) {
+					throw new Exception("Table " + describeTableBean.getTablename() + " is not existed");
+				}
+				List<String> listColumns = new ArrayList<String>();
+				List<String> listDataType = new ArrayList<String>();
+				List<TableColumn> list = tableinfo.getColumns();
+				for (TableColumn col : list) {
+					listColumns.add(col.getColumnName());
+					listDataType.add(col.getDataType());
+				}
+				
+				bean.setColumnNames(listColumns);
+				bean.setDataTypes(listDataType);
+				bean.setStatus(MessageInfo.SUCCESS);
 			}
-			List<String> listColumns = new ArrayList<String>();
-			List<String> listDataType = new ArrayList<String>();
-			List<TableColumn> list = tableinfo.getColumns();
-			for (TableColumn col : list) {
-				listColumns.add(col.getColumnName());
-				listDataType.add(col.getDataType());
-			}
-			
-			bean.setColumnNames(listColumns);
-			bean.setDataTypes(listDataType);
-			bean.setStatus(bean.SUCCESS);
 		}catch(Exception e) {
-			bean.setStatus(bean.ERROR);
+			bean.setStatus(MessageInfo.ERROR);
 			bean.setException("Describe table " + describeTableBean.getTablename() + " failed : " + e.getMessage());
 			e.printStackTrace();
 		}
